refactor: Changed file structure for rings
